requirements.txt
Flask-SQLAlchemy
Flask-Script
PyMySQL
